backend/requirements.txt
Django==4.2.11
python-dotenv==1.0.1
Requests==2.31.0
django-cors-headers==4.3.1